+++ src/main/java/com/kingpixel/cobblenotify/events/cobblemon/CatchEvent.java
@@ -3,6 +3,7 @@ package com.kingpixel.cobblenotify.events.cobblemon;
 import com.cobblemon.mod.common.api.Priority;
 import com.cobblemon.mod.common.api.events.CobblemonEvents;
 import com.cobblemon.mod.common.pokemon.Pokemon;
+import com.kingpixel.cobblenotify.CobbleNotify;
 import com.kingpixel.cobblenotify.config.Notifications;
 import com.kingpixel.cobblenotify.models.Notification;
 import net.minecraft.server.network.ServerPlayerEntity;
@@ -16,7 +17,9 @@ public class CatchEvent {
       try {
         Pokemon pokemon = evt.getPokemon();
         ServerPlayerEntity player = evt.getPlayer();
+        if (CobbleNotify.config.getWorldFilter().isBlackListed(player.getServerWorld())) return;
         for (Notification notification : Notifications.NOTIFICATIONS) {
+          if (notification.getWorldFilter().isBlackListed(player.getServerWorld())) continue;
           if (notification.computeCaught(pokemon, player)) return;
         }
       } catch (Exception e) {