+++ src/main/java/com/kingpixel/cobblenotify/events/cobblemon/DefeatEvent.java
@@ -2,6 +2,7 @@ package com.kingpixel.cobblenotify.events.cobblemon;
 
 import com.cobblemon.mod.common.api.Priority;
 import com.cobblemon.mod.common.api.events.CobblemonEvents;
+import com.kingpixel.cobblenotify.CobbleNotify;
 import com.kingpixel.cobblenotify.config.Notifications;
 import com.kingpixel.cobblenotify.models.Notification;
 
@@ -16,7 +17,9 @@ public class DefeatEvent {
         if (!battle.isPvW()) return;
         var killed = evt.getKilled().getOriginalPokemon();
         var killer = evt.getBattle().getPlayers().getFirst();
+        if (CobbleNotify.config.getWorldFilter().isBlackListed(killer.getServerWorld())) return;
         for (Notification notification : Notifications.NOTIFICATIONS) {
+          if (notification.getWorldFilter().isBlackListed(killer.getServerWorld())) continue;
           if (notification.computeDefeat(killed, killer)) return;
         }
       } catch (Exception e) {