+++ src/main/java/com/kingpixel/cobblenotify/events/cobblemon/TradeEvent.java
@@ -7,6 +7,7 @@ import com.kingpixel.cobblenotify.CobbleNotify;
 import com.kingpixel.cobblenotify.config.Notifications;
 import com.kingpixel.cobblenotify.models.Notification;
 import net.minecraft.server.network.ServerPlayerEntity;
+import net.minecraft.server.world.ServerWorld;
 
 import java.util.UUID;
 
@@ -19,10 +20,13 @@ public class TradeEvent {
       try {
         Pokemon pokemon1 = evt.getTradeParticipant1Pokemon();
         Pokemon pokemon2 = evt.getTradeParticipant2Pokemon();
-        UUID player1 = evt.getTradeParticipant1().getUuid();
-        UUID player2 = evt.getTradeParticipant2().getUuid();
+        ServerPlayerEntity player1 = getPlayerByUUID(evt.getTradeParticipant1().getUuid());
+        ServerPlayerEntity player2 = getPlayerByUUID(evt.getTradeParticipant2().getUuid());
+        ServerWorld world = player1.getServerWorld();
+        if (CobbleNotify.config.getWorldFilter().isBlackListed(world)) return;
         for (Notification notification : Notifications.NOTIFICATIONS) {
-          if (notification.computeTrade(pokemon1, pokemon2, getPlayerByUUID(player1), getPlayerByUUID(player2))) return;
+          if (notification.getWorldFilter().isBlackListed(world)) continue;
+          if (notification.computeTrade(pokemon1, pokemon2, player1, player2)) return;
         }
       } catch (Exception e) {
         e.printStackTrace();